Separate extensions into their own folder Give proper namespaces for extensions
