+++ Stratis.Bitcoin.Features.Consensus.Tests/TimeSyncBehaviorTest.cs
@@ -3,6 +3,7 @@ using Stratis.Bitcoin.Base;
 using Stratis.Bitcoin.Utilities;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Net;
 using Xunit;
