+++ Stratis.Bitcoin.Features.RPC/Controllers/FullNodeController.cs
@@ -13,6 +13,7 @@ using Stratis.Bitcoin.Interfaces;
 using Stratis.Bitcoin.Utilities;
 using Microsoft.Extensions.Logging;
 using Stratis.Bitcoin.Features.Consensus.CoinViews;
+using NBitcoin.Protocol;
 
 namespace Stratis.Bitcoin.Features.RPC.Controllers
 {
