+++ Stratis.Bitcoin/Extensions/ArgsExtensions.cs
@@ -1,0 +1,32 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace System
+{
+
+    /// <summary>
+    /// Extension methods for arguments array.
+    /// </summary>
+    public static class ArgsExtensions
+    {
+        /// <summary>
+        /// Obtains a value of command line argument.
+        /// <para>
+        /// It is expected that arguments are written on command line as <c>argName=argValue</c>, 
+        /// where argName usually (but does not need to) starts with "-".
+        /// </para>
+        /// <para>
+        /// The argValue can be wrapped with '"' quotes from both sides, in which case the quotes are removed, 
+        /// but it is not allowed for argValue to contain '"' inside the actual value.
+        /// </para>
+        /// </summary>
+        /// <param name="args">Application command line arguments.</param>
+        /// <param name="arg">Name of the command line argument which value should be obtained.</param>
+        /// <returns>Value of the specified argument or null if no such argument is found among the given list of arguments.</returns>
+        public static string GetValueOf(this string[] args, string arg)
+        {
+            return args.Where(a => a.StartsWith($"{arg}=")).Select(a => a.Substring($"{arg}=".Length).Replace("\"", "")).FirstOrDefault();
+        }
+    }
+}
